Rename SelfAttention to MultiHeadDotProductAttention.
This clarifies that the module can also be used for encoder-decoder
attention by specifying both inputs_q and inputs_kv.

Includes SelfAttention as a shim which hardcodes inputs_kv to None.
But in the future it probably makes sense to refactor these modules
so that the details that are only relevant for SelfAttention (e.g.
causal_mask and cache) are only exposed in that class.