make jax_utils.replicate match pmap devices
Currently jax_utils.replicate creates a ShardedDeviceArray with devices in the order of `jax.local_devices()`, but this is not always the same as the device assignment used for pmaps (it's the same for multi-host setups, but not for single-host, where we use the default XLA device assignment). Having it match also in the single-host case will avoid unnecessary device-to-device copies to re-shard the array when launching the pmap.