Don't automatically attach "template" module args.
"template" module constructor arguments are modules with parent=None
i.e. unbound modules.  Before we attached these automatically and
immediately during __init__, but the logic was problematic and its
unclear if this is even the correct thing to do as it removes choice
from the user about where to bind a module template inside their
modules.

Also, we're planning to clarify the templating api very soon, so
this should go for now.