Implement multi-scope transforms for linen.
The functional core api can lift functions that take multiple scopes.
We use this to handle transforms that act on modules that have had
other modules bound to outside scopes passed into them via dataclass
constructor arguments.  e.g. a shared embedding table passed in to
different parts of the network.

At the moment we don't check for or handle aliasing across the outer
module scopes.  We plan to fix this soon.